requirements.txt
fastapi
SQLAlchemy
uvicorn
pydantic
passlib
bcrypt
python-jose
python-multipart